piloting-sample-ai-predictor-main/requirements.txt
attrs==21.4.0
beautifulsoup4==4.10.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.11
cycler==0.11.0
filelock==3.4.1
gdown==4.3.0
google-auth==2.6.0
google-auth-oauthlib==0.4.6
idna==3.3
importlib-metadata==4.8.3
iniconfig==1.1.1
kiwisolver==1.3.1
matplotlib==3.3.4
numpy==1.19.5
oauthlib==3.2.0
opencv-python==4.5.5.62
packaging==21.3
pandas==1.1.5
pika==1.2.0
Pillow==8.4.0
pluggy==1.0.0
protobuf==3.19.4
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.7
PySocks==1.7.1
pytest==7.0.1
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scipy==1.5.4
seaborn==0.11.2
six==1.16.0
soupsieve==2.3.1
thop==0.0.31.post2005241907
tomli==1.2.3
torch==1.10.2
torchvision==0.11.3
tqdm==4.62.3
typing_extensions==4.0.1
urllib3==1.26.8
Werkzeug==2.0.3
zipp==3.6.0
